Add RouteTable propagation for AWS Transit Gateway
Previously, the CIDR route was added manually based on the
connection entry from the CSP Connector in the database. The CSP
Connector collected that address from the GCP VPC during creating
a connection since the Cloud Router and VPN Gateway require a
Network anyway.

This solution was invalid for two reasons:

* the VPC which holds Cloud Router and VPN Gateway on the GCP side
    is known as peering VPC - actual workload VPCs which we may
    want to connect will have their own CIDR addresses and this
    solution did not verify it

* the CSP Connector establishes a BGP session so all the routes
    are learned dynamically. The reason it was not picked by the
    AWS side was that Transit Gateway used for testing had no
    RoutePropagation enabled.

The new change adds updating the Transit Gateway picked as the AWS
Gateway to enable Route Propagation if it was disabled.

The change also includes some refactoring around CSP Connector.